requirements.txt
mesa>=3.0.0
numpy
pyyaml
solara
networkx
matplotlib
altair